requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
Django==4.2.5
django-crispy-forms==2.0
django-environ==0.11.2
django-htmx==1.9.0
Pillow==10.0.1
psycopg2-binary==2.9.7
sqlparse==0.4.4
typing-extensions==4.8.0
